split out shipping details to files
