+++ src/components/ShippingDetails/ArrivalDates.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+import { spacingSmall } from '../../styles/designTokens';
+
+const ArrivalDates = styled.span`
+  font-weight: bold;
+  margin-left: ${spacingSmall};
+`;
+
+export default ArrivalDates;
