+++ src/components/ShippingDetails/Message.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+import { spacingSmall } from '../../styles/designTokens';
+
+const Message = styled.div`
+  flex-grow: 1;
+  padding: 0 ${spacingSmall};
+`;
+
+export default Message;
