+++ src/components/ShippingDetails/Section.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const Section = styled.section`
+  margin: 0 auto;
+  width: 100rem;
+`;
+
+export default Section;
