+++ src/components/ShippingDetails/Wrapper.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const Wrapper = styled.div`
+  display: flex;
+  align-items: center;
+`;
+
+export default Wrapper;
